reorganized main/java
