+++ src/main/java/ingredients/IngredientsResponse.java
@@ -1,4 +1,4 @@
-package order.serializer;
+package ingredients;
 
 import java.util.List;
